+++ RobustRotaryEncoder.h
@@ -34,8 +34,8 @@ class RobustRotaryEncoder {
 
   void              begin(void);
   void              begin(uint8_t pinA, uint8_t pinB);
-  uint8_t           read(void);
-  uint8_t           incrementValue(void);
+  int               read(void);
+  int               incrementValue(void);
   bool              isChanged(void);
   bool              isCW(void);
   bool              isCCW(void);
@@ -43,7 +43,7 @@ class RobustRotaryEncoder {
  protected:
   uint8_t           m_pinA;     ///< encoder pin A
   uint8_t           m_pinB;     ///< encoder pin B
-  uint8_t           m_val;      ///< current increment value: 1 (CW), -1 (CCW), 0 (invalid)
+  int               m_val;      ///< current increment value: 1 (CW), -1 (CCW), 0 (invalid)
   uint8_t           m_code;     ///< previous / next code pattern
   uint16_t          m_store;    ///< stored pattern
 };
